requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
exceptiongroup==1.2.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.0
google==3.0.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.4
googleapis-common-protos==1.63.2
googletrans==4.0.0rc1
grpcio==1.65.0
grpcio-status==1.65.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
proto-plus==1.24.0
protobuf==5.27.2
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.20.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
selenium==4.22.0
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sympy==1.12.1
TgCrypto==1.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
trio==0.26.0
trio-websocket==0.11.1
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
